refactor : rabbitmq 관련 설정 한 디렉터리로 합침
