+++ src/main/java/com/jungle/navigation/rabbitmq/config/RMQProperties.java
@@ -1,4 +1,4 @@
-package com.jungle.navigation.alarm.config;
+package com.jungle.navigation.rabbitmq.config;
 
 public interface RMQProperties {
 	String WAITING_EXCHANGE_NAME = "exchange.waiting";
@@ -8,6 +8,10 @@ public interface RMQProperties {
 	String ALARM_EXCHANGE_NAME = "exchange.alarm";
 	String ALARM_ROUTING_KEY = "key.alarm.queue";
 	String ALARM_QUEUE_NAME = "alarm.queue";
+
+	String CHAT_EXCHANGE_NAME = "exchange.chat";
+	String CHAT_ROUTING_KEY = "key.chat.queue";
+	String CHAT_QUEUE_NAME = "chat.queue";
 
 	String DEAD_LETTER_EXCHANGE_NAME = "exchange.dead";
 	String DEAD_LETTER_ROUTING_KEY = "key.dead.queue";
