+++ src/main/java/com/jungle/navigation/rabbitmq/config/queue/DeadLetterQueueConfig.java
@@ -1,8 +1,8 @@
-package com.jungle.navigation.alarm.config.queue;
+package com.jungle.navigation.rabbitmq.config.queue;
 
-import static com.jungle.navigation.alarm.config.RMQProperties.DEAD_LETTER_EXCHANGE_NAME;
-import static com.jungle.navigation.alarm.config.RMQProperties.DEAD_LETTER_QUEUE_NAME;
-import static com.jungle.navigation.alarm.config.RMQProperties.DEAD_LETTER_ROUTING_KEY;
+import static com.jungle.navigation.rabbitmq.config.RMQProperties.DEAD_LETTER_EXCHANGE_NAME;
+import static com.jungle.navigation.rabbitmq.config.RMQProperties.DEAD_LETTER_QUEUE_NAME;
+import static com.jungle.navigation.rabbitmq.config.RMQProperties.DEAD_LETTER_ROUTING_KEY;
 
 import org.springframework.amqp.core.Binding;
 import org.springframework.amqp.core.BindingBuilder;
