+++ src/main/java/com/jungle/navigation/rabbitmq/config/queue/WaitingQueueConfig.java
@@ -1,10 +1,10 @@
-package com.jungle.navigation.alarm.config.queue;
+package com.jungle.navigation.rabbitmq.config.queue;
 
-import static com.jungle.navigation.alarm.config.RMQProperties.ALARM_EXCHANGE_NAME;
-import static com.jungle.navigation.alarm.config.RMQProperties.ALARM_ROUTING_KEY;
-import static com.jungle.navigation.alarm.config.RMQProperties.WAITING_EXCHANGE_NAME;
-import static com.jungle.navigation.alarm.config.RMQProperties.WAITING_QUEUE_NAME;
-import static com.jungle.navigation.alarm.config.RMQProperties.WAITING_ROUTING_KEY;
+import static com.jungle.navigation.rabbitmq.config.RMQProperties.ALARM_EXCHANGE_NAME;
+import static com.jungle.navigation.rabbitmq.config.RMQProperties.ALARM_ROUTING_KEY;
+import static com.jungle.navigation.rabbitmq.config.RMQProperties.WAITING_EXCHANGE_NAME;
+import static com.jungle.navigation.rabbitmq.config.RMQProperties.WAITING_QUEUE_NAME;
+import static com.jungle.navigation.rabbitmq.config.RMQProperties.WAITING_ROUTING_KEY;
 
 import org.springframework.amqp.core.Binding;
 import org.springframework.amqp.core.BindingBuilder;
